+++ src/app/components/invoice-pdf-template/invoice-items-table.tsx
@@ -1,4 +1,4 @@
-import { Text, View } from "@react-pdf/renderer";
+import { Text, View } from "@react-pdf/renderer/lib/react-pdf.browser";
 import { styles } from ".";
 import type { InvoiceData } from "@/app/schema";
 import { TRANSLATIONS } from "@/app/schema/translations";
