+++ src/lib/check-local-storage.ts
@@ -1,3 +1,5 @@
+"use client";
+
 /**
  * https://developer.mozilla.org/en-US/docs/Web/API/Web_Storage_API/Using_the_Web_Storage_API#testing_for_availability
  */
